+++ snake_web/constants/DSnakeWeb.py
@@ -2,5 +2,5 @@ from typing import Final
 
 
 class DSnakeWeb:
-    VERSION: Final[str] = "0.2.7"
+    VERSION: Final[str] = "0.2.8"
     POLL_INTERVAL: Final[int] = 300